allow replace secret value on project settings
